feat(chart): adding config and accessor methods
I've added accessor and config method to the base as convienence methods
for setting default config options and then  allowing the user to change
them as desired. I also added an accessor method as a convienence with
dealing with how to access  the data. We want our widgets to be as
modular as possible.